Klienci/jjj/product_service/product_service/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-cors-headers==4.7.0
django-nested-admin==4.1.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
